Relocated context files out of utils and into core
