+++ src/core/axioms.c
@@ -4,8 +4,8 @@
 #include <math.h>
 
 #include "axioms.h"
-#include "core/utils/context/context.h"
-#include "core/utils/context/environment.h"
+#include "core/context/context.h"
+#include "core/context/environment.h"
 #include "core/utils/log.h"
 #include "core/utils/types.h"
