+++ src/core/context/environment.c
