+++ src/core/context/environment.h
