+++ src/core/execution/execute.c
@@ -2,8 +2,8 @@
 #include <stdio.h>
 
 #include "execute.h"
-#include "core/utils/context/context.h"
-#include "core/utils/context/environment.h"
+#include "core/context/context.h"
+#include "core/context/environment.h"
 #include "core/utils/log.h"
 #include "core/utils/types.h"
