+++ src/core/parsing/codegen/ast.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 
 #include "ast.h"
-#include "core/utils/context/context.h"
+#include "core/context/context.h"
 #include "core/utils/log.h"
 #include "core/parsing/parserUtils.h"
 #include "core/utils/types.h"
