+++ src/core/parsing/parser.c
@@ -3,9 +3,9 @@
 #include <string.h>
 
 #include "parser.h"
-#include "core/utils/context/context.h"
+#include "core/context/context.h"
+#include "core/context/environment.h"
 #include "core/utils/log.h"
-#include "core/utils/context/environment.h"
 
 #include "core/parsing/codegen/tokenizer.h"
 #include "core/parsing/codegen/lexer.h"
