+++ src/core/utils/log.c
@@ -2,7 +2,7 @@
 #include <stdarg.h>
 
 #include "log.h"
-#include "core/utils/context/context.h"
+#include "core/context/context.h"
 
 static const int DEFAULT_BUFFER = 128;
